Use EventType.
