+++ data/serializer/src/test/java/tech/pegasys/teku/api/response/v1/EventTypeTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beaconrestapi.handlers.v1.events;
+package tech.pegasys.teku.api.response.v1;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
